chore: address review
  - Separate exported workflow tags into individual files
  - Add jsdoc compliant docstrings for all touched functions
