+++ src/js/samples/components/Tag/Icon.tsx
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+/**
+ * Styled component for the icon in a workflow tag.
+ *
+ *@Returns {StyledComponent}
+ */
+export const SampleItemLabelIcon = styled.span`
+    margin-right: 3px;
+    width: 12px;
+`;
